LayoutTests/imported/w3c/web-platform-tests/html/browsers/windows/browsing-context-expected.txt


FAIL Check that browsing context has new, ready HTML document assert_equals: The document's encoding should be 'UTF-8'. expected "UTF-8" but got "windows-1252"
PASS Check that new document nodes extant, empty 
PASS Check the document properties corresponding to the creator browsing context 
FAIL Check the history.length of the created browsing context assert_equals: The history.length should be 1. expected 1 but got 100
